added Profit / Loss. Cleaned up Code + comments
added a simple function to show profit loss on tab2 for closed positions. also has it where if the profit is negative the background color is red, positive gains will be green. Cleaned up the code to look a bit more professional and added more comments to explain the code.